Properly set the words in the popup in case of multiple tabs.
When two tabs had matching words, it would forget about the words in the
first tab. Store all results in a dictionary and retrieve them based on
the tabID.

Also fixed the scoping, now the background page has a single function
which returns the active set of words for the current tab.