Fewer false positives.
In some cases a false positive was generated. For example:

<ul><li>Foo</li><li>Bar</li></ul>

The textContent attribute would squash this together, resulting in
FooBar. To work around this, split a word when it has a capital letter
inside. This may not work in all cases, but at least it works with
websites with proper capatilization.